Move benchmark suites to /suites
