+++ benchmarks/suites/add.js
@@ -1,16 +1,16 @@
 'use strict';
 const { Big: BigJs } = require('big.js');
-const { Big } = require('../lib/bigdecimal.js');
-const { BigDecimal: GWTDecimal } = require('bigdecimal');
+const { Big } = require('../../lib/bigdecimal.js');
 const { BigNumber } = require('bignumber.js');
-const Decimal = require('decimal.js');
+const { BigDecimal: GWTDecimal } = require('bigdecimal');
 const Benchmark = require('benchmark');
-const { bigDecimals, bigDecimalsBigjs, bigDecimalsBigNumber, bigDecimalsDecimal, bigDecimalsGWT } = require('./test_numbers');
-const { attachEventsAndRun } = require('./utils.js');
+const { bigDecimals, bigDecimalsBigjs, bigDecimalsBigNumber, bigDecimalsDecimal, bigDecimalsGWT } = require('../test_numbers');
+const Decimal = require('decimal.js');
+const { attachEventsAndRun } = require('../utils.js');
 
-const suite = new Benchmark.Suite('Subtract');
+const suite = new Benchmark.Suite('Add');
 
-const initialValue = '123033';
+const initialValue = '0';
 
 let resBigDec = Big(initialValue);
 let resBigJs = BigJs(initialValue);
@@ -20,23 +20,23 @@ let resGWT = GWTDecimal(initialValue);
 
 suite.add('Bigdecimal.js', function () {
     for (const x of bigDecimals) {
-        resBigDec = resBigDec.subtract(x);
+        resBigDec = resBigDec.add(x);
     }
 }).add('Big.js', function () {
     for (const x of bigDecimalsBigjs) {
-        resBigJs = resBigJs.sub(x);
+        resBigJs = resBigJs.add(x);
     }
 }).add('BigNumber.js', function () {
     for (const x of bigDecimalsBigNumber) {
-        resBigNumJs = resBigNumJs.minus(x);
+        resBigNumJs = resBigNumJs.plus(x);
     }
 }).add('decimal.js', function () {
     for (const x of bigDecimalsDecimal) {
-        resDecimalJs = resDecimalJs.sub(x);
+        resDecimalJs = resDecimalJs.plus(x);
     }
 }).add('GWTBased', function () {
     for (const x of bigDecimalsGWT) {
-        resGWT = resGWT.subtract(x);
+        resGWT = resGWT.add(x);
     }
 });
