+++ benchmarks/suites/integer_round.js
@@ -1,29 +1,30 @@
 'use strict';
 const Benchmark = require('benchmark');
-const { bigDecimals, bigDecimalsBigjs, bigDecimalsBigNumber, bigDecimalsDecimal, bigDecimalsGWT } = require('./test_numbers');
-const { attachEventsAndRun } = require('./utils.js');
+const { RoundingMode } = require('../../lib/bigdecimal.js');
+const { bigDecimals, bigDecimalsBigjs, bigDecimalsBigNumber, bigDecimalsDecimal, bigDecimalsGWT } = require('../test_numbers');
+const { attachEventsAndRun } = require('../utils.js');
 
-const suite = new Benchmark.Suite('Abs');
+const suite = new Benchmark.Suite('Integer rounding');
 
 suite.add('Bigdecimal.js', function () {
     for (const x of bigDecimals) {
-        x.abs();
+        x.setScale(0, RoundingMode.HALF_UP);
     }
 }).add('Big.js', function () {
     for (const x of bigDecimalsBigjs) {
-        x.abs();
+        x.round(0, 1);
     }
 }).add('BigNumber.js', function () {
     for (const x of bigDecimalsBigNumber) {
-        x.abs();
+        x.integerValue();
     }
 }).add('decimal.js', function () {
     for (const x of bigDecimalsDecimal) {
-        x.abs();
+        x.round();
     }
 }).add('GWTBased', function () {
     for (const x of bigDecimalsGWT) {
-        x.abs();
+        x.setScale(0, RoundingMode.HALF_UP);
     }
 });
