+++ benchmarks/suites/positive_pow.js
@@ -1,29 +1,32 @@
 'use strict';
 const Benchmark = require('benchmark');
-const { bigDecimals, bigDecimalsBigjs, bigDecimalsBigNumber, bigDecimalsDecimal, bigDecimalsGWT } = require('./test_numbers');
-const { attachEventsAndRun } = require('./utils.js');
+const { bigDecimals, bigDecimalsBigjs, bigDecimalsBigNumber, bigDecimalsDecimal, bigDecimalsGWT } = require('../test_numbers');
+const { attachEventsAndRun } = require('../utils.js');
 
-const suite = new Benchmark.Suite('Multiply');
+const suite = new Benchmark.Suite('Positive pow');
+
+const exponents = [0, 1, 2, 10, 99];
+const length = exponents.length;
 
 suite.add('Bigdecimal.js', function () {
     for (let i = 0; i < bigDecimals.length - 1; i++) {
-        bigDecimals[i].multiply(bigDecimals[i+1]);
+        bigDecimals[i].pow(exponents[i%length]);
     }
 }).add('Big.js', function () {
     for (let i = 0; i < bigDecimalsBigjs.length - 1; i++) {
-        bigDecimalsBigjs[i].mul(bigDecimalsBigjs[i+1]);
+        bigDecimalsBigjs[i].pow(exponents[i%length]);
     }
 }).add('BigNumber.js', function () {
     for (let i = 0; i < bigDecimalsBigNumber.length - 1; i++) {
-        bigDecimalsBigNumber[i].multipliedBy(bigDecimalsBigjs[i+1]);
+        bigDecimalsBigNumber[i].pow(exponents[i%length]);
     }
 }).add('decimal.js', function () {
     for (let i = 0; i < bigDecimalsDecimal.length - 1; i++) {
-        bigDecimalsDecimal[i].times(bigDecimalsDecimal[i+1]);
+        bigDecimalsDecimal[i].pow(exponents[i%length]);
     }
 }).add('GWTBased', function () {
     for (let i = 0; i < bigDecimalsGWT.length - 1; i++) {
-        bigDecimalsGWT[i].multiply(bigDecimalsGWT[i+1]);
+        bigDecimalsGWT[i].pow(exponents[i%length]);
     }
 });
