+++ benchmarks/suites/subtract.js
@@ -1,16 +1,16 @@
 'use strict';
 const { Big: BigJs } = require('big.js');
-const { Big } = require('../lib/bigdecimal.js');
-const { BigNumber } = require('bignumber.js');
+const { Big } = require('../../lib/bigdecimal.js');
 const { BigDecimal: GWTDecimal } = require('bigdecimal');
-const Benchmark = require('benchmark');
-const { bigDecimals, bigDecimalsBigjs, bigDecimalsBigNumber, bigDecimalsDecimal, bigDecimalsGWT } = require('./test_numbers');
+const { BigNumber } = require('bignumber.js');
 const Decimal = require('decimal.js');
-const { attachEventsAndRun } = require('./utils.js');
+const Benchmark = require('benchmark');
+const { bigDecimals, bigDecimalsBigjs, bigDecimalsBigNumber, bigDecimalsDecimal, bigDecimalsGWT } = require('../test_numbers');
+const { attachEventsAndRun } = require('../utils.js');
 
-const suite = new Benchmark.Suite('Add');
+const suite = new Benchmark.Suite('Subtract');
 
-const initialValue = '0';
+const initialValue = '123033';
 
 let resBigDec = Big(initialValue);
 let resBigJs = BigJs(initialValue);
@@ -20,23 +20,23 @@ let resGWT = GWTDecimal(initialValue);
 
 suite.add('Bigdecimal.js', function () {
     for (const x of bigDecimals) {
-        resBigDec = resBigDec.add(x);
+        resBigDec = resBigDec.subtract(x);
     }
 }).add('Big.js', function () {
     for (const x of bigDecimalsBigjs) {
-        resBigJs = resBigJs.add(x);
+        resBigJs = resBigJs.sub(x);
     }
 }).add('BigNumber.js', function () {
     for (const x of bigDecimalsBigNumber) {
-        resBigNumJs = resBigNumJs.plus(x);
+        resBigNumJs = resBigNumJs.minus(x);
     }
 }).add('decimal.js', function () {
     for (const x of bigDecimalsDecimal) {
-        resDecimalJs = resDecimalJs.plus(x);
+        resDecimalJs = resDecimalJs.sub(x);
     }
 }).add('GWTBased', function () {
     for (const x of bigDecimalsGWT) {
-        resGWT = resGWT.add(x);
+        resGWT = resGWT.subtract(x);
     }
 });
